Move Web Api Metrics config class to Web module
- The Web Api Metric configuration class has a dependency to Servlet
Filter which causes some tests to fail
