+++ dhis-2/dhis-services/dhis-service-schema/pom.xml
@@ -39,12 +39,6 @@
       <artifactId>hamcrest-library</artifactId>
     </dependency>
 
-    <dependency>
-      <groupId>javax.servlet</groupId>
-      <artifactId>javax.servlet-api</artifactId>
-      <scope>test</scope>
-    </dependency>
-
   </dependencies>
   <properties>
     <rootDir>../../</rootDir>
