+++ dhis-2/dhis-support/dhis-support-system/src/main/java/org/hisp/dhis/monitoring/metrics/HibernateMetricsConfig.java
@@ -98,7 +98,7 @@ public class HibernateMetricsConfig
         extends MetricsEnabler
     {
         @Override
-        ConfigurationKey getConfigKey()
+        protected ConfigurationKey getConfigKey()
         {
             return MONITORING_HIBERNATE_ENABLED;
         }
