+++ dhis-2/dhis-support/dhis-support-system/src/main/java/org/hisp/dhis/monitoring/metrics/JvmMetricsConfig.java
@@ -88,7 +88,7 @@ public class JvmMetricsConfig
         MetricsEnabler
     {
         @Override
-        ConfigurationKey getConfigKey()
+        protected ConfigurationKey getConfigKey()
         {
             return MONITORING_JVM_ENABLED;
         }
