+++ dhis-2/dhis-support/dhis-support-system/src/main/java/org/hisp/dhis/monitoring/metrics/MetricsEnabler.java
@@ -51,5 +51,5 @@ public abstract class MetricsEnabler
         return !isTestRun( conditionContext ) && getBooleanValue( getConfigKey() );
     }
 
-    abstract ConfigurationKey getConfigKey();
+    protected abstract ConfigurationKey getConfigKey();
 }
